rename files for Rails 6.0
