+++ lib/view_source_map_rails_6_0.rb
